@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -31,6 +31,7 @@ dependencies:
   flutter_bloc: ^7.0.0
   snapping_sheet: ^3.0.0+2
   equatable: ^2.0.0
+  sqflite: ^1.3.0
 
 
   # The following adds the Cupertino Icons font to your application.
